update: #686: `Address::from`の定義を`src/parser/parse.rs`に移動
